atelier_ideal/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-filebrowser-no-grappelli==3.8.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
